compatible with upcoming beast 2.5

+++ src/speciesnetwork/NetworkStatLogger.java
@@ -36,7 +36,7 @@ public class NetworkStatLogger extends CalculationNode implements Loggable, Func
     }
 
     @Override
-    public void log(int sample, PrintStream out) {
+    public void log(long sample, PrintStream out) {
         final Network speciesNetwork = speciesNetworkInput.get();
         if (logHeightInput.get()) {
             // network root height
